Refactored position.string_to_move() to uci implementation
